Fix multiple entries in result
There was a problem when multiple entries for the same word, e.g.:
searching for word "cat" in dictionaryapi returns multiple results with
all words with the same name as "cat".

This problem is an issue, because when you import your csv file into
Anki, Anki will see each new "cat" and overwrite the existing one,
making that the only the last one is saved in the card.